+++ tests/Unit/Report/ComponentGenerator/TableAggregateComponentTest.php
@@ -8,14 +8,23 @@ use PhpBench\Expression\Ast\StringNode;
 use PhpBench\Report\Bridge\ExpressionBridge;
 use PhpBench\Report\ComponentGenerator\TableAggregateComponent;
 use PhpBench\Report\ComponentGeneratorInterface;
+use PhpBench\Report\ComponentGenerator\TableAggregate\ExpandColumnProcessor;
+use PhpBench\Report\ComponentGenerator\TableAggregate\ExpressionColumnProcessor;
+use PhpBench\Report\ComponentGenerator\TableAggregate\GroupHelper;
 use PhpBench\Report\Model\Builder\TableBuilder;
 use PhpBench\Report\Model\Table;
+use PhpBench\Report\Model\TableColumnGroup;
 
 class TableAggregateComponentTest extends ComponentGeneratorTestCase
 {
     public function createGenerator(): ComponentGeneratorInterface
     {
-        return new TableAggregateComponent($this->container()->get(ExpressionBridge::class));
+        $evaluator = $this->container()->get(ExpressionBridge::class);
+
+        return new TableAggregateComponent($evaluator, [
+            'expression' => new ExpressionColumnProcessor($evaluator),
+            'expand' => new ExpandColumnProcessor($evaluator),
+        ]);
     }
 
     public function testNoConfiguration(): void
@@ -111,6 +120,69 @@ class TableAggregateComponentTest extends ComponentGeneratorTestCase
                 'hello' => new ListNode([new StringNode('goodbye')]),
                 'goodbye' => new ListNode([new StringNode('hello'), new StringNode('goodbye')]),
             ]
-        ])->build(), $table);
+        ])->addGroups([new TableColumnGroup(GroupHelper::DEFAULT_GROUP_NAME, 2)])->build(), $table);
+    }
+
+    public function testExpandKey(): void
+    {
+        $frame = DataFrame::fromRowSeries([
+            ['bench1', 'subject1', 1],
+            ['bench1', 'subject2', 1],
+            ['bench2', 'subject1', 1],
+            ['bench2', 'subject2', 1],
+        ], ['name', 'subject', 'time']);
+
+        $table = $this->generate($frame, [
+            TableAggregateComponent::PARAM_ROW => [
+                'one' => 'first(partition["subject"])',
+                'by_time' => [
+                    'type' => 'expand',
+                    'partition' => 'name',
+                    'cols' => [
+                        '{{ key }}' => 'mode(partition["time"])',
+                    ]
+                ]
+            ],
+            TableAggregateComponent::PARAM_PARTITION => ['subject'],
+        ]);
+        assert($table instanceof Table);
+        self::assertCount(2, $table->rows());
+        self::assertCount(3, $table->columnNames());
+        self::assertEquals(['one', 'bench1', 'bench2'], $table->columnNames());
+    }
+
+    public function testGroups(): void
+    {
+        $frame = DataFrame::fromRowSeries([
+            ['bench1', 'subject1', 1],
+            ['bench1', 'subject2', 1],
+            ['bench2', 'subject1', 1],
+            ['bench2', 'subject2', 1],
+        ], ['name', 'subject', 'time']);
+
+        $table = $this->generate($frame, [
+            'groups' => [
+                'time' => [
+                    'cols' => ['by_time'],
+                    'label' => 'Time',
+                ],
+            ],
+            TableAggregateComponent::PARAM_ROW => [
+                'one' => 'first(partition["subject"])',
+                'by_time' => [
+                    'type' => 'expand',
+                    'partition' => 'name',
+                    'cols' => [
+                        '{{ key }}' => 'mode(partition["time"])',
+                    ]
+                ]
+            ],
+            TableAggregateComponent::PARAM_PARTITION => ['subject'],
+        ]);
+        assert($table instanceof Table);
+        self::assertCount(2, $table->rows());
+        self::assertCount(3, $table->columnNames());
+        self::assertEquals(2, count($table->columnGroups()));
+        self::assertEquals(['one', 'bench1', 'bench2'], $table->columnNames());
     }
 }